collector/requirements.txt
pyserial>=3.5
pip>=22.0.2
